+++ src/scriptlets/scriptlets-list.js
@@ -59,3 +59,4 @@ export * from './trusted-set-constant';
 export * from './inject-css-in-shadow-dom';
 export * from './remove-node-text';
 export * from './trusted-replace-node-text';
+export * from './evaldata-prune';